Rename: BaseEntity, UserEntity 위치 변경
