+++ user-api/src/main/java/com/biengual/userapi/user/domain/UserReader.java
@@ -1,7 +1,5 @@
 package com.biengual.userapi.user.domain;
 
-import com.biengual.userapi.user.domain.entity.UserEntity;
-
 public interface UserReader {
     UserEntity findUser(Long userId, String email);
 }
