+++ user-api/src/main/java/com/biengual/userapi/user/domain/UserService.java
@@ -1,7 +1,6 @@
 package com.biengual.userapi.user.domain;
 
 import com.biengual.userapi.oauth2.domain.info.OAuth2UserPrincipal;
-import com.biengual.userapi.user.domain.entity.UserEntity;
 import com.biengual.userapi.user.presentation.UserResponseDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
