+++ user-api/src/main/java/com/biengual/userapi/user/infrastructure/UserReaderImpl.java
@@ -3,7 +3,7 @@ package com.biengual.userapi.user.infrastructure;
 import com.biengual.userapi.annotation.DataProvider;
 import com.biengual.userapi.message.error.exception.CommonException;
 import com.biengual.userapi.user.domain.UserReader;
-import com.biengual.userapi.user.domain.entity.UserEntity;
+import com.biengual.userapi.user.domain.UserEntity;
 import com.biengual.userapi.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
