+++ user-api/src/main/java/com/biengual/userapi/user/infrastructure/UserStoreImpl.java
@@ -4,7 +4,7 @@ import com.biengual.userapi.annotation.DataProvider;
 import com.biengual.userapi.message.error.exception.CommonException;
 import com.biengual.userapi.user.domain.UserCommand;
 import com.biengual.userapi.user.domain.UserStore;
-import com.biengual.userapi.user.domain.entity.UserEntity;
+import com.biengual.userapi.user.domain.UserEntity;
 import com.biengual.userapi.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
