+++ user-api/src/main/java/com/biengual/userapi/user/repository/UserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.biengual.userapi.user.domain.entity.UserEntity;
+import com.biengual.userapi.user.domain.UserEntity;
 
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
