+++ user-api/src/test/java/com/biengual/userapi/service/UserServiceTest.java
@@ -19,7 +19,7 @@ import org.mockito.MockitoAnnotations;
 import com.biengual.userapi.exception.user.UserNotFoundException;
 import com.biengual.userapi.user.presentation.UserRequestDto;
 import com.biengual.userapi.user.presentation.UserResponseDto;
-import com.biengual.userapi.user.domain.entity.UserEntity;
+import com.biengual.userapi.user.domain.UserEntity;
 import com.biengual.userapi.user.domain.enums.Gender;
 import com.biengual.userapi.user.domain.enums.UserStatus;
 import com.biengual.userapi.user.repository.UserRepository;
